@@ -60,46 +60,57 @@ public class EthereumRPC {
         
     // Swift4 warning bug - https://bugs.swift.org/browse/SR-6265
    // static func execute<T: Encodable, U: Decodable>(session: URLSession, url: URL, method: String, params: T, receive: U.Type, id: Int = 1, completion: @escaping ((Error?, JSONRPCResult<U>?) -> Void)) -> Void {
+    @available(*, deprecated, message: "Prefer async alternative instead")
     public static func execute<T: Encodable, U: Decodable>(session: URLSession, url: URL, method: String, params: T, receive: U.Type, id: Int = 1, completion: @escaping ((Error?, Any?) -> Void)) -> Void {
+        async {
+            do {
+                let result: Any? = try await execute(session: session, url: url, method: method, params: params, receive: receive, id: id)
+                completion(nil, result)
+            } catch {
+                completion(error, nil)
+            }
+        }
+    }
+    
+    
+    public static func execute<T: Encodable, U: Decodable>(session: URLSession, url: URL, method: String, params: T, receive: U.Type, id: Int = 1) async throws -> Any {
         
         if type(of: params) == [Any].self {
             // If params are passed in with Array<Any> and not caught, runtime fatal error
-            completion(JSONRPCError.encodingError, nil)
-            return
+            throw JSONRPCError.encodingError
         }
-
+        
         var request = URLRequest(url: url, cachePolicy: .reloadIgnoringLocalCacheData)
         request.httpMethod = "POST"
         request.addValue("application/json", forHTTPHeaderField: "Content-Type")
         request.addValue("application/json", forHTTPHeaderField: "Accept")
         
         let rpcRequest = JSONRPCRequest(jsonrpc: "2.0", method: method, params: params, id: id)
-        guard let encoded = try? JSONEncoder().encode(rpcRequest) else {
-            completion(JSONRPCError.encodingError, nil)
-            return
-        }
+        let encoded = try JSONEncoder().encode(rpcRequest)
         request.httpBody = encoded
-        
-        let task = session.dataTask(with: request) { (data, response, error) in
-            if let data = data {
-                if let result = try? JSONDecoder().decode(JSONRPCResult<U>.self, from: data) {
-                    return completion(nil, result.result)
-                } else if let result = try? JSONDecoder().decode([JSONRPCResult<U>].self, from: data) {
-                    let resultObjects = result.map{ return $0.result }
-                    return completion(nil, resultObjects)
-                } else if let errorResult = try? JSONDecoder().decode(JSONRPCErrorResult.self, from: data) {
-                    print("Ethereum response error: \(errorResult.error)")
-                    return completion(JSONRPCError.executionError(errorResult), nil)
-                } else if let response = response as? HTTPURLResponse, response.statusCode < 200 || response.statusCode > 299 {
-                    return completion(JSONRPCError.requestRejected(data), nil)
-                } else {
-                    return completion(JSONRPCError.noResult, nil)
+                
+        return try await withCheckedThrowingContinuation { continuation in
+            let task = session.dataTask(with: request) { (data, response, error) in
+                if let data = data {
+                    if let result = try? JSONDecoder().decode(JSONRPCResult<U>.self, from: data) {
+                        return continuation.resume(returning: (result.result))
+                    } else if let result = try? JSONDecoder().decode([JSONRPCResult<U>].self, from: data) {
+                        let resultObjects = result.map{ return $0.result }
+                        return continuation.resume(returning: (resultObjects))
+                    } else if let errorResult = try? JSONDecoder().decode(JSONRPCErrorResult.self, from: data) {
+                        print("Ethereum response error: \(errorResult.error)")
+                        return continuation.resume(throwing: JSONRPCError.executionError(errorResult))
+                    } else if let response = response as? HTTPURLResponse, response.statusCode < 200 || response.statusCode > 299 {
+                        return continuation.resume(throwing: JSONRPCError.requestRejected(data))
+                    } else {
+                        return continuation.resume(throwing: JSONRPCError.noResult)
+                    }
                 }
+                
+                continuation.resume(throwing: JSONRPCError.unknownError)
             }
             
-            completion(JSONRPCError.unknownError, nil)
+            task.resume()
         }
-        
-        task.resume()
     }
 }
